% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions-code.R
\name{Cquantile}
\alias{Cquantile}
\title{Fast computation of quantiles}
\usage{
Cquantile(xx, p)
}
\arguments{
\item{xx}{Numeric vector we are obtaining quantiles for.}

\item{p}{Numeric vector of quantiles.}
}
\value{
Numeric vector containing quantiles, possibly greater than one.
}
\description{
Helper function which takes as input a vector and obtains quantiles for it.
Number of quantiles may be greater than one.
}

