\name{perccal-package}
\alias{perccal-package}
\alias{perccal}
\docType{package}
\title{
Computing Confidence Intervals Via Double Bootstrap 'perc-cal' Method 
}
\description{
  
Contains functions which allow the user to compute confidence intervals 
quickly using the double bootstrap-based percentile calibrated 
('perc-cal') method for linear regression coefficients. 

}
\details{
  \tabular{ll}{
    Package: \tab perccal\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2016-06-14\cr
    License: \tab GPL-3\cr
    }

	Contains functions which allow users to compute confidence intervals 
	quickly using the double bootstrap-based percentile calibrated 
	('perc-cal') method for linear regression coefficients. 
	
	The help of Justin Bleich is strongly acknowledged.

}
\author{
  Daniel McCarthy

  Maintainer: Daniel McCarthy <danielmc@wharton.upenn.edu>
}
\references{

Efron, Bradley; Tibshirani, Robert J. \dQuote{An Introduction to the Bootstrap}
1994. Book. Publisher: CRC Press.

McCarthy, Daniel; Zhang, Kai; Berk, Richard; Brown, Lawrence; Buja, Andreas;
George, Edward; Zhao, Linda. \dQuote{Calibrated Percentile Double Bootstrap 
For Robust Linear Regression Inference} 2016. Available on arXiv: 
https://arxiv.org/abs/1511.00273

}

\keyword{ package }

