% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions-code.R
\name{sample_rcpp}
\alias{sample_rcpp}
\title{Sample from [1,2,...,N] with replacement nsamp times in Rcpp.}
\usage{
sample_rcpp(N, nsamp)
}
\arguments{
\item{N}{Largest integer to sample from.}

\item{nsamp}{number of samples from [1,2,...,N] with replacement to obtain.}
}
\value{
samps nsamp-length vector of samples from [1,2,...,N] with replacement to obtain.
}
\description{
Helper function which samples from [1,2,...,N] with replacement
nsamp times in Rcpp.
}

